requirements.txt
flask==3.1.0
google-generativeai==0.8.5
google-api-python-client==2.166.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
python-dotenv==1.1.0
duckduckgo-search
gunicorn
